signout api and beautify admin dashboard

+++ admin-app/src/actions/auth.actions.js
@@ -48,8 +48,19 @@ export const isUserLoggedIn = () => async (dispatch) => {
 }
 
 export const signout = () => async(dispatch) => {
-    localStorage.clear()
-    dispatch({
-        type: authConstants.LOGOUT_REQUEST
-    })
+
+    dispatch({ type: authConstants.LOGOUT_REQUEST })
+    const res = await axios.post('/admin/signout')
+
+    if (res.status === 200) {
+        localStorage.clear()
+        dispatch({ type: authConstants.LOGOUT_SUCCESS })
+    }
+    else {
+        dispatch({ 
+            type: authConstants.LOGOUT_FAILURE,
+            payload: { error: res.data.error }
+        })
+    }
+
 }
